Move keyword and range vespa search tests into separate file
